Updated according to format archetype
