+++ nexus-repository-r/src/main/java/org/sonatype/nexus/repository/r/internal/RRecipeSupport.groovy
@@ -23,6 +23,7 @@ import org.sonatype.nexus.repository.http.PartialFetchHandler
 import org.sonatype.nexus.repository.httpclient.HttpClientFacet
 import org.sonatype.nexus.repository.r.RFacet
 import org.sonatype.nexus.repository.r.RRestoreFacet
+import org.sonatype.nexus.repository.r.internal.security.RSecurityFacet
 import org.sonatype.nexus.repository.routing.RoutingRuleHandler
 import org.sonatype.nexus.repository.search.SearchFacet
 import org.sonatype.nexus.repository.security.SecurityHandler
