+++ nexus-repository-r/src/main/java/org/sonatype/nexus/repository/r/internal/security/RFormatSecurityContributor.java
@@ -10,13 +10,14 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.r.internal;
+package org.sonatype.nexus.repository.r.internal.security;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
 import org.sonatype.nexus.repository.Format;
+import org.sonatype.nexus.repository.r.internal.RFormat;
 import org.sonatype.nexus.repository.security.RepositoryFormatSecurityContributor;
 
 /**
